+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2026-08-30
+
+### Added
+- `User` Pydantic models
+- In-memory seed data
+- `UserService` service layer - domain-driven design CRUD.
+- API layer CRUD with DI (`get_user_service`)
+- Password hashing via `bcrypt`
+- UserRole change guard on 'update user' requires admin role
+- Exception handler in `main.py`
+- Unit tests for `UserService`
+- Component tests for full user module API path 
+
 ## [0.1.0] - 2026-08-23
 
 ### Added